refactor(api): handle pagination and sort in getAllItems API

+++ ApartmentManagement/src/main/java/com/ou/dto/response/PaginationResponse.java
@@ -12,7 +12,7 @@ import java.util.List;
 @NoArgsConstructor
 @FieldDefaults(level = AccessLevel.PRIVATE)
 public class PaginationResponse<T> {
-    int count;
+    Long count;
     LinkResponse links;
     List<T> results;
 }
